category/games/2026/02/20260209-072051-vector-columns-falling-match/pyproject.toml
[project]
name = "vector-columns-falling-match"
version = "1.0.0"
description = "Align three or more matching gems in vertical, horizontal, or diagonal lines as they fall"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = []
